+++ NCexamples/qtransport/transmission/scripts/Makefile
@@ -6,13 +6,14 @@ cp2nc:
 
 
 clean3:
-	rm -r   3.scatter+tbtrans/0*
+	rm -r   3.scatter_tbtrans/0*
 
 clean4:
 	rm -r 	4.post_processing/files 4.post_processing/result* 4.post_processing/output.yaml
 
 clean:
+	rm      lap*
 	rm -r 	1.elec/Au*/OUT 
-	rm 		2.model/*fdf
-	rm -r 	3.scatter+tbtrans/0*
+	rm      2.model/*fdf
+	rm -r 	3.scatter_tbtrans/0*
 	rm -r 	4.post_processing/files 4.post_processing/result* 4.post_processing/output.yaml
