+++ __init__.py
@@ -1,6 +1,7 @@
 # External packages
 '''
-    Written by JPark 2021.10.20: Refactorizing modules
+    Written by JP 2021.10.20: Refactorizing modules\n
+    Merged by JP  2023.09.14: Quantum Transport by SHY
 '''
 from __future__ import print_function
