+++ bin/catalytic_info.py
