+++ bin/fdf2xcrysden.py
@@ -4,7 +4,7 @@ from nanocore import siesta as s2
 from nanocore.aux import vis
 
 def fdf2xcrysden(fdf_name):
-    at = s2.read_fdf(fdf_name)
+    at = s2.read_struct(fdf_name)
     vis.show_xcrysden(at)
 
 if __name__ == '__main__':
