+++ solvers/__init__.py
@@ -2,3 +2,8 @@
 subpackage solvers
 each subpackage works independently
 """
+from . import catalysis
+#from . import phonon_qt
+from . import qttransport
+
+__all__ = ['catalysis', 'qttransport']
